Addressed PR review comment
