+++ test/e2e/shoot/upgrade_multi_zone.go
@@ -20,18 +20,18 @@ import (
 
 	"github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	"github.com/gardener/gardener/test/e2e"
-	shoothaupdatesuite "github.com/gardener/gardener/test/utils/shoots/update/ha"
+	"github.com/gardener/gardener/test/utils/shoots/update/highavailability"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("Shoot Tests", Label("Shoot", "ha", "multi-zone"), func() {
+var _ = Describe("Shoot Tests", Label("Shoot", "high-availability", "multi-zone"), func() {
 	f := defaultShootCreationFramework()
 	f.Shoot = e2e.DefaultShoot("e2e-upgrade-mz")
 	f.Shoot.Spec.ControlPlane = nil
 
-	It("Create, Upgrade (control plane failTolearanceType to 'zone')  and Delete Shoot", func() {
+	It("Create, Upgrade (non-HA to HA with failure tolerance type 'zone')  and Delete Shoot", func() {
 		By("Create Shoot")
 		ctx, cancel := context.WithTimeout(parentCtx, 15*time.Minute)
 		defer cancel()
@@ -39,7 +39,10 @@ var _ = Describe("Shoot Tests", Label("Shoot", "ha", "multi-zone"), func() {
 		Expect(f.CreateShootAndWaitForCreation(ctx, false)).To(Succeed())
 		f.Verify()
 
-		shoothaupdatesuite.RunTest(ctx, f.ShootFramework, v1beta1.FailureToleranceTypeZone)
+		By("Upgrade Shoot (non-HA to HA with failure tolerance type 'zone')")
+		ctx, cancel = context.WithTimeout(parentCtx, 15*time.Minute)
+		defer cancel()
+		highavailability.UpdateAndVerify(ctx, f.ShootFramework, v1beta1.FailureToleranceTypeZone)
 
 		By("Delete Shoot")
 		ctx, cancel = context.WithTimeout(parentCtx, 15*time.Minute)
