+++ test/utils/shoots/update/highavailability/update.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package ha
+package highavailability
 
 import (
 	"context"
@@ -25,31 +25,33 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	appsv1 "k8s.io/api/apps/v1"
+	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/utils/pointer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-// RunTest runs the HA control-plane upgrade tests for an existing shoot cluster.
-func RunTest(
+// UpdateAndVerify runs the HA control-plane update tests for an existing shoot cluster.
+func UpdateAndVerify(
 	ctx context.Context,
 	f *framework.ShootFramework,
-	haType gardencorev1beta1.FailureToleranceType,
+	failureToleranceType gardencorev1beta1.FailureToleranceType,
 
 ) {
-	By("updating shoot's control plane failTolearanceType to " + string(haType))
+	By("Update Shoot control plane to HA with failure tolerance type " + string(failureToleranceType))
 	Expect(f.UpdateShoot(ctx, func(shoot *gardencorev1beta1.Shoot) error {
 		shoot.Spec.ControlPlane = &gardencorev1beta1.ControlPlane{
 			HighAvailability: &gardencorev1beta1.HighAvailability{
 				FailureTolerance: gardencorev1beta1.FailureTolerance{
-					Type: haType,
+					Type: failureToleranceType,
 				},
 			},
 		}
 		return nil
 	})).To(Succeed())
 
-	By("verifying shoot's control plane components for failTolearanceType: " + string(haType))
+	By("Verify Shoot's control plane components")
 	verifyTSC(ctx, f.SeedClient, f.Shoot, f.ShootSeedNamespace())
 	verifyEtcdAffinity(ctx, f.SeedClient, f.Shoot, f.ShootSeedNamespace())
 }
@@ -143,4 +145,78 @@ func verifyEtcdAffinity(ctx context.Context, seedClient kubernetes.Interface, sh
 		Expect(c.Get(ctx, client.ObjectKeyFromObject(sts), sts)).To(Succeed())
 		Expect(sts.Spec.Template.Spec.Affinity).Should(BeEquivalentTo(affinity))
 	}
+}
+
+// DeployZeroDownTimeValidatorJob deploys k8s job in cluster, which ensures
+// zero down time by continuously checking kube-apiserver health.
+// This job fails once health check fails and associated pod results in error status.
+func DeployZeroDownTimeValidatorJob(ctx context.Context, c client.Client, testName, namespace, token string) (*batchv1.Job, error) {
+	job := batchv1.Job{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "zero-down-time-validator-" + testName,
+			Namespace: namespace,
+		},
+		Spec: batchv1.JobSpec{
+			Template: corev1.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: namespace,
+					Labels: map[string]string{
+						"networking.gardener.cloud/to-dns":             "allowed",
+						"networking.gardener.cloud/to-shoot-apiserver": "allowed",
+					},
+				},
+				Spec: corev1.PodSpec{
+					Containers: []corev1.Container{
+						{
+							Name:    "validator",
+							Image:   "alpine/curl",
+							Command: []string{"/bin/sh"},
+
+							//To avoid flakiness, consider downtime when curl fails consecutively back-to-back.
+							Args: []string{"-ec",
+								"echo '" +
+									"failed=0 ; threshold=2 ; " +
+									"while [ $failed -lt $threshold ] ; do  " +
+									"$(curl -k https://kube-apiserver/healthz -H \"Authorization: " + token + "\" -s -f  -o /dev/null ); " +
+									"if [ $? -gt 0 ] ; then let failed++; echo \"etcd is unhealthy and retrying\"; continue;  fi ; " +
+									"echo \"kube-apiserver is healthy\";  touch /tmp/healthy; let failed=0; " +
+									"sleep 1; done;  echo \"kube-apiserver is unhealthy\"; exit 1;" +
+									"' > test.sh && sh test.sh",
+							},
+							ReadinessProbe: &corev1.Probe{
+								InitialDelaySeconds: int32(5),
+								FailureThreshold:    int32(2),
+								PeriodSeconds:       int32(1),
+								SuccessThreshold:    int32(3),
+								ProbeHandler: corev1.ProbeHandler{
+									Exec: &corev1.ExecAction{
+										Command: []string{
+											"cat",
+											"/tmp/healthy",
+										},
+									},
+								},
+							},
+							LivenessProbe: &corev1.Probe{
+								InitialDelaySeconds: int32(5),
+								FailureThreshold:    int32(2),
+								PeriodSeconds:       int32(1),
+								ProbeHandler: corev1.ProbeHandler{
+									Exec: &corev1.ExecAction{
+										Command: []string{
+											"cat",
+											"/tmp/healthy",
+										},
+									},
+								},
+							},
+						},
+					},
+					RestartPolicy: corev1.RestartPolicyNever,
+				},
+			},
+			BackoffLimit: pointer.Int32(0),
+		},
+	}
+	return &job, c.Create(ctx, &job)
 }
